requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.9.1
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
asyncssh==2.18.0
atpublic==5.0
attrs==24.2.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.22.6
dvc==3.58.0
dvc-data==3.16.7
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.1.0
filelock==3.16.1
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
funcy==2.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grandalf==0.8
grpcio==1.68.1
gto==1.7.2
h5py==3.12.1
hydra-core==1.3.2
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
iterative-telemetry==0.0.9
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.13.1
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opt_einsum==3.4.0
orjson==3.10.12
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydot==3.0.3
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scmrepo==3.3.9
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sqltrie==0.11.1
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.5.0
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.14.0
typing_extensions==4.5.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zc.lockfile==3.0.post1
